// packages/nx-docusaurus/src/generators/application/schema.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface DocusaurusGeneratorExecutor {
  /**
   * The name of the application.
   */
  name: string;
  /**
   * The directory of the new application.
   */
  directory?: string;
  /**
   * Skip formatting files.
   */
  skipFormat?: boolean;
  /**
   * Add tags to the project (used for linting)
   */
  tags?: string;
  /**
   * Split the project configuration into <projectRoot>/project.json rather than including it inside workspace.json.
   */
  standaloneConfig?: boolean;
}